% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_closest_noaa_station}
\alias{get_closest_noaa_station}
\title{Find the closest NOAA station}
\usage{
get_closest_noaa_station(.date_start, .date_end, lat, long)
}
\arguments{
\item{.date_start}{the starting date that the NOAA station should contain data for}

\item{.date_end}{the ending date that the NOAA station should contain data for}

\item{lat}{a double representing latitude}

\item{long}{a double representing longitude}
}
\value{
string containing the NOAA station ID
}
\description{
Finds the closest NOAA station based on a given latitude, longitude. Only considers NOAA stations that contain temperature, precipitation, and wind speed for the provided dates and are within +/- 1 degree.
}
\examples{
# set_api_key_noaa('<token>')
# date_start <- '2021-09-10'
# date_end <- '2021-09-22'
# lat <- 40.7812
# long <- -73.9665
# get_closest_noaa_station(date_start, date_end, lat, long)
}
