% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_openweather_forecast}
\alias{get_openweather_forecast}
\title{Retrieve forecasted weather data from the OpenWeather API}
\usage{
get_openweather_forecast()
}
\value{
a tibble of weather data
}
\description{
Returns the 7-day temperature (F), precipitation (T/F), and wind speed (mph) forecast for Central Park, NY. Precipitation defined as >= 0.35 forecasted probability of rain. Must set API token via set_api_key_openweather() prior to use. Request token here get token here https://openweathermap.org/full-price#current.
}
\examples{
# set_api_key_openweather('<token>')
get_openweather_forecast()
}
\references{
https://openweathermap.org/api/one-call-api
}
