% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key.R
\name{set_api_key_noaa}
\alias{set_api_key_noaa}
\title{Set API key for NOAA}
\usage{
set_api_key_noaa(api_key)
}
\arguments{
\item{api_key}{the api key string}
}
\description{
Set your NOAA API key. This stores it as an environment variable so it can be used with simpleweather functions. You can install the key to your .Renviron so calling this function is not required for each R session.
}
\details{
You can obtain an API key for free here: https://www.ncdc.noaa.gov/cdo-web/webservices/v2
}
\examples{
\dontrun{
set_api_key_noaa("<key>")
}
}
