% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_noaa}
\alias{get_noaa}
\title{Retrieve historical weather data from the NOAA API}
\usage{
get_noaa(.date_start, .date_end)
}
\arguments{
\item{date_start}{the starting date to retrieve data for}

\item{date_end}{the end date to retrieve data for}
}
\value{
a data frame
}
\description{
Returns daily temperature (F), precipitation (T/F), or wind (mph) data for Central Park, NY using the NOAA API. Must set API token via Sys.setenv(token_noaa = '<token>') prior to use. Request token here get token here https://www.ncdc.noaa.gov/cdo-web/token.
}
\examples{
# set_api_key_noaa('<token>')
get_noaa('2021-09-10', '2021-09-22')
}
\references{
https://www.ncdc.noaa.gov/cdo-web/webservices/v2
}
