% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_openweather_historical}
\alias{get_openweather_historical}
\title{Retrieve last 5 days weather data from the OpenWeather API}
\usage{
get_openweather_historical()
}
\value{

}
\description{
Returns the last 5 days temperature (F), precipitation (T/F), and wind speed (mph) for Central Park, NY. Must set API token via Sys.setenv(token_openweather = '<token>') prior to use. Request token here get token here https://openweathermap.org/full-price#current.
}
\examples{
# set_api_key_openweather('<token>')
get_openweather_historical()
}
\references{
https://openweathermap.org/api/one-call-api
}
