% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_weather}
\alias{get_weather}
\title{Retrieve historical or forecasted weather}
\usage{
get_weather(.dates)
}
\arguments{
\item{.dates}{a vector of dates}
}
\value{
a tibble of weather data with nrows == length(.dates)
}
\description{
Returns historical or forecasted temperature (F), precipitation (T/F), and wind speed (mph) for Central Park, NY.
}
\examples{
# set_api_key_noaa('<token>')
# set_api_key_openweather('<token>')
dates <- seq(Sys.Date() - 10, Sys.Date() + 5, by = 'day')
get_weather(.dates = dates)
}
